[UI] Переработано отображение пользователя и неавторизованного пользователя Resolve #183
