+++ src/frontend/app/(root)/page.tsx
@@ -7,3 +7,4 @@ export default function Home() {
     </section>
   );
 }
+
