+++ src/frontend/app/(root)/wishlists/[uuid]/page.tsx
