+++ src/frontend/app/layout.tsx
@@ -10,6 +10,7 @@ import Menu from "@/components/menu";
 import { UserItem } from "@/components/user";
 import { fontSans } from "@/config/fonts";
 import { siteConfig } from "@/config/site";
+import Login from "@/components/login";
 
 export const metadata: Metadata = {
   title: {
@@ -34,9 +35,6 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const cookieStore = await cookies();
-  const isLogin = cookieStore.has("access_token");
-
   return (
     <html suppressHydrationWarning lang="en">
       <head title="Мои желания">
@@ -48,34 +46,19 @@ export default async function RootLayout({
       <body
         className={clsx(
           "h-screen bg-background font-sans antialiased",
-          fontSans.variable,
+          fontSans.variable
         )}
       >
         <Providers themeProps={{ attribute: "class", defaultTheme: "dark" }}>
-          {!isLogin ? (
-            children
-          ) : (
-            <div className="grid grid-cols-12 gap-4">
-              <div className="col-span-full md:col-span-3 lg:col-span-2 py-10 flex-col">
-                <div className="mx-auto text-center">
-                  <UserItem />
-                </div>
-                <div className="mt-10">
-                  <Menu />
-                </div>
-              </div>
-              <Toaster
-                position="bottom-right"
-                toastOptions={{
-                  className: "!bg-content1 dark:!text-white !text-base-content",
-                }}
-              />
-              <div className="col-span-full md:col-span-9 lg:col-span-10 py-10 px-2">
-                <main>{children}</main>
-                <footer className="w-full flex items-center justify-center py-3" />
-              </div>
-            </div>
-          )}
+          <>
+            {children}
+            <Toaster
+              position="bottom-right"
+              toastOptions={{
+                className: "!bg-content1 dark:!text-white !text-base-content",
+              }}
+            />
+          </>
         </Providers>
       </body>
     </html>
