+++ src/frontend/components/wishlistDetail.tsx
@@ -54,7 +54,7 @@ const Wishes = observer(({ wishlistUUID }: { wishlistUUID: string }) => {
     setItems(
       items.filter((value) => {
         return value.uuid !== wish.uuid;
-      }),
+      })
     );
   }
 
@@ -70,7 +70,7 @@ const Wishes = observer(({ wishlistUUID }: { wishlistUUID: string }) => {
   ));
 
   return (
-    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4 p-2">
+    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4 px-4">
       <div className="col-span-full px-4">
         <WishlistDetail wishlist={wishlist} />
       </div>
